refactor(package): move database into data
