+++ app/src/main/java/com/example/ezvault/authentication/authentication/FirebaseAuthenticationStrategy.java
@@ -1,8 +1,8 @@
 package com.example.ezvault.authentication.authentication;
 
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.model.User;
-import com.example.ezvault.database.UserService;
+import com.example.ezvault.data.database.UserService;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
