+++ app/src/main/java/com/example/ezvault/view/EditItemDetails.java
@@ -42,9 +42,9 @@ import com.example.ezvault.viewmodel.SelectedItemViewModel;
 import com.example.ezvault.view.adapter.PhotoAdapter;
 import com.example.ezvault.R;
 import com.example.ezvault.camera.GalleryAction;
-import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.database.ImageDAO;
-import com.example.ezvault.database.ItemDAO;
+import com.example.ezvault.data.database.FirebaseBundle;
+import com.example.ezvault.data.database.ImageDAO;
+import com.example.ezvault.data.database.ItemDAO;
 import com.example.ezvault.model.Image;
 import com.example.ezvault.model.Item;
 import com.example.ezvault.model.SerialPrediction;
