+++ app/src/main/java/com/example/ezvault/viewmodel/LoginFragment.java
@@ -18,7 +18,7 @@ import androidx.navigation.Navigation;
 import com.example.ezvault.R;
 import com.example.ezvault.authentication.authentication.AuthenticationHandler;
 import com.example.ezvault.authentication.authentication.EmailPasswordAuthenticationStrategy;
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.FragmentUtils;
 import com.example.ezvault.utils.UserManager;
