+++ app/src/main/java/com/example/ezvault/viewmodel/NewUserFragment.java
@@ -19,7 +19,7 @@ import com.example.ezvault.R;
 import com.example.ezvault.authentication.registration.EmailPasswordRegistrationStrategy;
 import com.example.ezvault.authentication.registration.RegistrationException;
 import com.example.ezvault.authentication.registration.RegistrationHandler;
-import com.example.ezvault.database.FirebaseBundle;
+import com.example.ezvault.data.database.FirebaseBundle;
 import com.example.ezvault.utils.textwatchers.MirroredTextWatcher;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.textwatchers.PasswordWatcher;
