+++ app/src/main/java/com/example/ezvault/viewmodel/ProfileFragment.java
@@ -15,9 +15,9 @@ import androidx.navigation.Navigation;
 
 import com.example.ezvault.R;
 import com.example.ezvault.authentication.registration.RegistrationException;
-import com.example.ezvault.database.FirebaseBundle;
-import com.example.ezvault.database.RawUserDAO;
-import com.example.ezvault.database.UserService;
+import com.example.ezvault.data.database.FirebaseBundle;
+import com.example.ezvault.data.database.RawUserDAO;
+import com.example.ezvault.data.database.UserService;
 import com.example.ezvault.model.ItemList;
 import com.example.ezvault.utils.textwatchers.NonEmptyTextWatcher;
 import com.example.ezvault.utils.textwatchers.PasswordWatcher;
